requirements.txt
logdecorator==2.4
python-escpos==2.2.0
